added some changes
