+++ apps/main/src/app/lib/Assets/SVG/SponsorUsAssets/Ticket.tsx
