+++ apps/main/src/app/lib/Assets/SVG/SponsorUsAssets/TicketBooth.tsx
