+++ apps/main/src/app/sponsor-us/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import React from "react";
-import TicketBooth from "../lib/Assets/SponsorUsAssets/TicketBooth";
-import Ticket from "../lib/Assets/SponsorUsAssets/Ticket";
+import TicketBooth from "../lib/Assets/SVG/SponsorUsAssets/TicketBooth";
+import Ticket from "../lib/Assets/SVG/SponsorUsAssets/Ticket";
 import Button from "@repo/ui/Button";
 import useDevice from "@repo/util/hooks/useDevice";
