+++ apps/main/src/app/team/Sections/Teams.tsx
@@ -5,6 +5,7 @@ import Image from "next/image";
 import LinkedInLogo from "@repo/ui/LinkedInLogo";
 import useWindowSize from "@repo/util/hooks/useWindowSize";
 import OurTeamDropdown from "../OurTeamDropdown";
+import Icon from "../components/Icon.tsx";
 
 type HeadshotProps = {
   name: string;
@@ -137,6 +138,15 @@ const Teams = () => {
           {getTeamsToDisplay().map(([teamName]) => (
             <TeamSections key={teamName} team={teamName} />
           ))}
+        </div>
+
+        <div>
+          <Icon
+            imageLocation="/headshots/emma.png"
+            name="Emma my queen"
+            size={150}
+            url="www.linkedin.com"
+          ></Icon>
         </div>
 
         <div className="hidden tablet:block">
